# src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API de Inscripciones Académicas",
    "description": "Documentación de la API REST del MP-S2",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/estudiantes/": {
      "post": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "Registrar un estudiante nuevo",
        "description": "",
        "responses": {
          "201": {
            "description": "201 Created",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 3
                },
                "nombre": {
                  "type": "string",
                  "example": "Juan"
                },
                "email": {
                  "type": "string",
                  "example": "juan123@gmail.com"
                },
                "bootcamp": {
                  "type": "string",
                  "example": "Backend"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "400 Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "400 Bad Request"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "Estudiante",
                "required": [
                  "nombre",
                  "email",
                  "bootcamp"
                ],
                "properties": {
                  "nombre": {
                    "type": "string",
                    "example": "Maria"
                  },
                  "email": {
                    "type": "string",
                    "example": "maria123@gmail.com"
                  },
                  "bootcamp": {
                    "type": "string",
                    "example": "FrontEnd"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "Listar estudiantes",
        "description": "",
        "parameters": [
          {
            "name": "bootcamp",
            "in": "query",
            "description": "Filtrar por bootcamp",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de estudiantes",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 3
                  },
                  "nombre": {
                    "type": "string",
                    "example": "Roberto"
                  },
                  "email": {
                    "type": "string",
                    "example": "rober123@gmail.com"
                  },
                  "bootcamp": {
                    "type": "string",
                    "example": "Frontend"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/estudiantes/{id}": {
      "put": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "Modifica un estudiante por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID numérico del producto"
          }
        ],
        "responses": {
          "200": {
            "description": "Estudiante modificado",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "nombre": {
                  "type": "string",
                  "example": "Maria"
                },
                "email": {
                  "type": "string",
                  "example": "maria456@gmail.com"
                },
                "bootcamp": {
                  "type": "string",
                  "example": "FullStack"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Estudiante no encontrado",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Estudiante no encontrado"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "Elimina un estudiante del registro",
        "description": "Elimina un estudiante del registro",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Estudiantes"
        ],
        "summary": "Obtener estudiante por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Estudiante encontrado",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "nombre": {
                  "type": "string",
                  "example": "Maria"
                },
                "email": {
                  "type": "string",
                  "example": "maria123@gmail.com"
                },
                "bootcamp": {
                  "type": "string",
                  "example": "Frontend"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "Estudiante no encontrado",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Estudiante no encontrado"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/api/estudiantes/api/status": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}